third_party/WebKit/LayoutTests/fast/animation/request-animation-frame-prefix-expected.txt

CONSOLE WARNING: line 18: 'webkitRequestAnimationFrame' is vendor-specific. Please use the standard 'requestAnimationFrame' instead.
Tests the timestamps provided to prefixed webkitRequestAnimationFrame callbacks

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS legacyFirstTimestamp is >= firstTimestamp
PASS legacySecondTimestamp is >= secondTimestamp
PASS deltaError < 0.001 is true
